app/requirements.txt
flask-appconfig>=0.10
flask-bootstrap
flask-nav
flask-bootstrap
flask-wtf